+++ assets/styles/app.scss
@@ -2,4 +2,4 @@
 @import "~flatpickr/dist/flatpickr";
 @import "~@fortawesome/fontawesome-free/scss/fontawesome.scss";
 @import "~@fortawesome/fontawesome-free/scss/solid.scss";
-@import "sidebar";
+@import "./sidebar.scss";
