+++ templates/base.html.twig
@@ -8,11 +8,12 @@
         {% block scripts %}{% endblock %}
     </head>
     <body>
-    <div id="container">
-    {{ include('layouts/slidebar.html.twig') }}
-    <div id="content">
-        {% block body %}{% endblock %}
-    </div>
-</div>
-</body>
+        <div id="container">
+            {{ include('layouts/sidebar.html.twig') }}
+            <div id="content">
+                {% block body %}{% endblock %}
+            </div>
+            {{ encore_entry_script_tags('app') }}
+        </div>
+    </body>
 </html>
